Little Endian Support!!!
This is awesome, I finally got Little Endian support working! Just like I thought it would be, the code itself wasn't that big of a change to implement, haha. The hard part was figuring out how to implement it. Wow! It took a long time to get to this point.

I found a few sources that really helped with the last stage of figuring it out:
https://www.reddit.com/r/MCPE/comments/aj54ev/how_to_read_world_save_data/eeto1ej/
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/getInt16

For that MDN reference, it was the endian parameter that made the lightbulb go off. I hadn't noticed how the tags were being parsed with the `getInt8` (and related) methods until today, which was a groundbreaking find! This is how I added the Little Endian support. Those methods allow you to pass a boolean for whether the data should be parsed as Little Endian or not.
I also added a parameter to the `write` modules which allows you to compile to NBT with GZIP compression, but it doesn't appear to be working correctly at the moment. Parsing from an existing GZIP-compressed NBT file still works as intended though.

I am ecstatic that this works now! I'm proud that I finally found the solution. Time to see where things can go now!