Re-Node Support!
Found and fixed issues with using the module in Node! Now things work great there too. Now that I've figured out Node-specific fixes, I have to merge the changes with the existing browser support, and use feature-detection for when specific parts should run in Node or not. Epic! (I could always just use a build tool for this, but naaah :P)