BigInt Support
Looked into the BigInt number type within JavaScript, and figured out how to better store long-type tags as BigInt once converted into a live JavaScript object. When stringifying to JSON, longs will now be a string of the number followed by the `n` character, which is a safe way to hold it's value in JSON without converting back and forth from Number to BigInt when opening with JavaScript.

A few articles really helped out again this time, here's what they were:
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/BigInt
https://dev.to/benlesh/bigint-and-json-stringify-json-parse-2m8p
https://stackoverflow.com/questions/53103695/how-to-read-64-bit-integer-from-an-arraybuffer-dataview-in-javascript

Other pages that helped with some of my testing, as well as some experimental implementations that I may use to convert NBT objects with types to plain JSON:
https://stackoverflow.com/questions/25783161/how-to-check-if-two-files-have-the-same-content
https://minecraft.fandom.com/wiki/NBT_format#:~:text=7%20External%20links-,SNBT%20format,-This%20feature%20is

I was working on this a few days ago, but I'm commiting it to the repo a little later since I had a few busy days with work, and Disneyland with the whole store :D