Experimental SNBT
This is an experimental few changes that I started working on to try and create my mixed SNBT-JSON syntax. It started to get complicated kind of quickly, and I wanted to end up rewriting other parts of the library instead of add this new features before the other parts were ready to support it.