Module Restructure
Starting to organize some of the files. This is one of the possibly more severe breaking ones so I did it first before the other changes come next.
I started working on this version on around Wednesday, and I'm starting to add them to the repo now to make it easier as I change more of the project structure.